Add dynamic_settings.yml
This contains Tag type definitions as well as definitions for the use pages. Because this can change between implementations, it's better to define them in initializers through yml files. Theoretically, we could put this in the DB instead, but us using a relational database means we can't add arbitrary fields to the YML without necessitating schema changes.